decouple speedo config from overlay
